Fix overlapping impls
